pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "coord2region"
dynamic = ["version"]  # Automatically fetch version from Git tags
authors = [
  { name = "Hamza Abdelhedi", email = "hamza.abdelhedii@gmail.com" }
]
description = "Find region name for a given MNI coordinate in a selected atlas"
readme = "README.md"
requires-python = ">=3.10,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
license = "BSD-3-Clause"

dependencies = [
    "biopython",
    "mne",
    "nibabel",
    "nilearn==0.12.0",
    "nimare==0.4.0", # pinned to match nilearn 0.12.0
    "numpy",
    "pandas",
    "pydantic>=2.6.0",
    "scipy",
    "requests",
    "pyyaml",
    "openpyxl",
    "Pillow",
    "fpdf",
    "anthropic",
    "diffusers",
    "google-genai",
    "openai>=1.0.0",
    "transformers",
    "pooch>=1.5",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest-cov",
    "ruff",
    "torch>=2.0.0",
]
docs = [
    "insegel",
    "myst-parser",
    "pydata-sphinx-theme",
    "sphinx>=4.0.0",
    "sphinx-autoapi",
    "sphinx-copybutton",
    "sphinx-gallery",
    "sphinxcontrib-mermaid",
]

[project.urls]
Homepage = "https://github.com/BabaSanfour/Coord2Region"
Issues = "https://github.com/BabaSanfour/Coord2Region/issues"

[project.scripts]
coord2region = "coord2region.cli:main"

[tool.setuptools_scm]
write_to = "coord2region/_version.py"

[tool.setuptools.packages.find]
include = ["coord2region", "coord2region.*"]
exclude = ["paper*", "docs*", "examples*", "tests*"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["docs", "examples", "tests"]

[tool.ruff.lint]
select = ["E", "F", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
